Revert "Autoload Types (#2209)"
This reverts commit 224402f0ba945017a391be5d9cd5bd273ce6593e.
