+++ lib/grape/validations/types/array_coercer.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'dry_type_coercer'
+
 module Grape
   module Validations
     module Types
