+++ lib/grape/validations/types/dry_type_coercer.rb
@@ -1,5 +1,14 @@
 # frozen_string_literal: true
 
+require 'dry-types'
+
+module DryTypes
+  # Call +Dry.Types()+ to add all registered types to +DryTypes+ which is
+  # a container in this case. Check documentation for more information
+  # https://dry-rb.org/gems/dry-types/1.2/getting-started/
+  include Dry.Types()
+end
+
 module Grape
   module Validations
     module Types
@@ -43,7 +52,7 @@ module Grape
         def initialize(type, strict = false)
           @type = type
           @strict = strict
-          @scope = strict ? Grape::DryTypes::Strict : Grape::DryTypes::Params
+          @scope = strict ? DryTypes::Strict : DryTypes::Params
         end
 
         # Coerces the given value to a type which was specified during
